+++ toyml/classification/naive_bayes/__init__.py
@@ -1,0 +1,7 @@
+from .gaussian_naive_bayes import GaussianNaiveBayes
+from .multinomial_naive_bayes import MultinomialNaiveBayes
+
+__all__ = [
+    "GaussianNaiveBayes",
+    "MultinomialNaiveBayes",
+]
